course/06-up-blog-news/requirements.txt
markdownify
openpyxl
psycopg2-binary
scrapy_redis
pandas
